change updating the friends steam info and owned games from updating on an interval to updating when requested on the stats page if a certain time has passed since the last update
