+++ controllers/api/index.js
@@ -1,7 +1,7 @@
 const router = require("express").Router();
 
-const userRoutes = require("./login.js");
+const userRoutes = require("./users.js");
 
-router.use("/", userRoutes);
+router.use("/users", userRoutes);
 
 module.exports = router;
