+++ public/js/login.js
@@ -5,12 +5,12 @@ const loginFormHandler = async (event) => {
   event.preventDefault();
   const username = document.querySelector("#username-input").value.trim();
   const password = document.querySelector("#password-input").value.trim();
-  console.log('you made it')
+  // console.log('you made it')
   if (username && password) {
     usernameInputForm.reset();
     passwordInputForm.reset();
 
-    const response = await fetch("/api/login", {
+    const response = await fetch("/api/users/login", {
       method: "POST",
       body: JSON.stringify({ username, password }),
       headers: { "Content-Type": "application/json" },
