+++ seeds/userData.js
@@ -7,6 +7,8 @@ const UserData = [{
   steam_avatar_full: 'https://avatars.akamai.steamstatic.com/5a9b5198d6858953b772c86a9e52a6234d1b64e8_full.jpg',
   steam_username: 'Goose, papa?',
   profile_url: "https://steamcommunity.com/profiles/76561198134108288/",
+  communityvisibilitystate: 3
+
 },
 {
   username: "david44",
@@ -15,6 +17,7 @@ const UserData = [{
   steam_avatar_full: 'https://avatars.akamai.steamstatic.com/5a9b5198d6858953b772c86a9e52a6234d1b64e8_full.jpg',
   steam_username: 'Goose, papa?',
   profile_url: "https://steamcommunity.com/profiles/76561198134108288/",
+  communityvisibilitystate: 3
 },
 {
   username: "davinchi",
@@ -23,6 +26,7 @@ const UserData = [{
   steam_avatar_full: 'https://avatars.akamai.steamstatic.com/9684c19be46eb8889669c0c791ddcb5c509838f6_full.jpg',
   steam_username: 'Rev. Richard Wayne Gary Wayne',
   profile_url: "https://steamcommunity.com/profiles/76561197961411971/",
+  communityvisibilitystate: 3
 },
 {
   username: "teslaMan",
@@ -31,6 +35,7 @@ const UserData = [{
   steam_avatar_full: 'https://avatars.akamai.steamstatic.com/fef49e7fa7e1997310d705b2a6158ff8dc1cdfeb_full.jpg',
   steam_username: 'davismasont',
   profile_url: "https://steamcommunity.com/profiles/76561199214467239/",
+  communityvisibilitystate: 3
 },
 {
   username: "Olo",
@@ -39,6 +44,7 @@ const UserData = [{
   steam_avatar_full: 'https://avatars.akamai.steamstatic.com/d84d56b37bb698158830696b3bef366e16777271_full.jpg',
   steam_username: 'OLO',
   profile_url: "https://steamcommunity.com/profiles/76561198006651484/",
+  communityvisibilitystate: 3
 },
 {
   username: "EdgarAllenBro",
@@ -46,7 +52,8 @@ const UserData = [{
   steam_id: "76561197960434622",
   steam_avatar_full: "https://technical.ly/wp-content/uploads/2013/11/EdgarAllanBro.jpg",
   steam_username: 'Edgar Allen Bro',
-  profile_url: "https://steamcommunity.com/profiles/76561197961411971/"
+  profile_url: "https://steamcommunity.com/profiles/76561197961411971/",
+  communityvisibilitystate: 3
 },
 {
   username: "mark",
@@ -54,7 +61,8 @@ const UserData = [{
   steam_id: "76561197972449074",
   steam_avatar_full: "https://avatars.akamai.steamstatic.com/14c9c3484d8df9255e977e3f95f0cb620bb2d26b_full.jpg",
   steam_username: 'Chairman Miao',
-  profile_url: "https://steamcommunity.com/profiles/76561197972449074/"
+  profile_url: "https://steamcommunity.com/profiles/76561197972449074/",
+  communityvisibilitystate: 3
 },
 {
   username: "roda",
@@ -62,7 +70,8 @@ const UserData = [{
   steam_id: "76561198216156851",
   steam_avatar_full: "https://avatars.akamai.steamstatic.com/7b58968bc4bac5edf9ce016150879e0c1539d8e2_full.jpg",
   steam_username: 'Payday',
-  profile_url: "https://steamcommunity.com/profiles/76561198216156851/"
+  profile_url: "https://steamcommunity.com/profiles/76561198216156851/",
+  communityvisibilitystate: 3
 },
 {
   username: "andrew",
@@ -70,7 +79,8 @@ const UserData = [{
   steam_id: "76561198153468156",
   steam_avatar_full: "https://avatars.akamai.steamstatic.com/ecfdbb7c31116c2cad7dfa1a8bf0f8a2abff018a_full.jpg",
   steam_username: 'Hex: Le GOOSE',
-  profile_url: "https://steamcommunity.com/profiles/76561198153468156/"
+  profile_url: "https://steamcommunity.com/profiles/76561198153468156/",
+  communityvisibilitystate: 3
 },
 {
   username: "teda",
@@ -78,7 +88,8 @@ const UserData = [{
   steam_id: "76561198195868989",
   steam_avatar_full: "https://avatars.akamai.steamstatic.com/c4a818da21ea23595d82759dc2f905204b1ea35d_full.jpg",
   steam_username: 'TedaToo$hort',
-  profile_url: "https://steamcommunity.com/id/TedaTooShort/"
+  profile_url: "https://steamcommunity.com/id/TedaTooShort/",
+  communityvisibilitystate: 3
 },
 {
   username: "pharce",
@@ -86,7 +97,8 @@ const UserData = [{
   steam_id: "76561197976402774",
   steam_avatar_full: "https://avatars.akamai.steamstatic.com/a51d41ae22e8d24839f62981cc22fa158581b642_full.jpg",
   steam_username: 'Pharce',
-  profile_url: "https://steamcommunity.com/profiles/76561197976402774/"
+  profile_url: "https://steamcommunity.com/profiles/76561197976402774/",
+  communityvisibilitystate: 3
 },
 ]
